refactor(editors): move Button sizing to a card's Layout tab
How big a launchpad's buttons are is a question about the card's layout,
not about what is on the buttons — and on the scaled style it is the
card's own size that answers it. So the Button sizing dropdown and
Buttons across slider move out of the three cards' Content editors and
into the card settings modal's Layout tab, under a Buttons heading right
below the card's own width and height.

The modal draws them once for every kind that declares the new
`tileButtons` flag (links, commands, templater) rather than each card
editor calling a shared helper, so `tileSizingSettings` folds into
`CardSettingsModal` and leaves cardbodies.ts, which no card module was
calling any more.

The fixed style's own Button size pixel slider stays where it has always
been, in the Content tab, and only shows while that style is on.
