+++ src/locales/zh.ts
@@ -1805,6 +1805,7 @@ export const zh: Translations = {
 				"输入框的粗细与卡片高度一致 — 在“排列”模式下拖动卡片边缘即可让搜索栏更粗或更细。",
 		},
 		tiles: {
+			heading: "按钮",
 			sizing: "按钮尺寸方式",
 			sizingDesc:
 				"按钮是随卡片一起缩放（就像卡片随面板缩放那样），还是无论卡片多大都保持固定的像素尺寸。" +
